Move test classes to correct packages
