+++ app/src/test/java/de/elanev/studip/android/app/data/model/MockServer.java
@@ -6,7 +6,7 @@
  * http://www.gnu.org/licenses/gpl.html
  */
 
-package de.elanev.studip.android.app.data.net.oauth;
+package de.elanev.studip.android.app.data.model;
 
 import de.elanev.studip.android.app.data.datamodel.Server;
